e2e: deploy the operator from the pipeline built image
Deploy the secondary scheduler operator from a PR built image.
This way any change made to the operator gets tested as part of a PR.

Note: IMAGE_FORMAT env is deprecated. Once the env is gone it can be
replaced by RELEASE_IMAGE_LATEST which contains the registry as well.